[gsupplicant] Cancel D-Bus calls in a deadlock-free way. JB#41874
The problem with g_cancellable_disconnect() is that it grabs a global
mutex. Since this library is not even supposed to be thread-safe, that
lock is useless to us and yet it may cause a deadlock, at least with
certain versions of glib. Let's use g_signal_handler_disconnect() to
disconnect from GCancellable and be happy.